requirements.txt
pygments
django
djangorestframework
django-filter
markdown
